+++ piri/collection_handlers.py
@@ -1,7 +1,8 @@
 from typing import Any, Dict, List, Union
 
-from returns.result import safe
+from returns.result import Failure, ResultE, Success, safe
 
+from piri.constants import ALIAS, PATH
 from piri.valuetypes import MapValue, ValueTypes
 
 
@@ -69,3 +70,42 @@ def fetch_list_by_keys(
         return collection  # type: ignore
 
     raise ValueError('Non list data found: ', str(collection))
+
+
+def iterable_data_handler(raw_data, paths) -> ResultE[list]:
+    """Iterate and create all combinations from list of paths."""
+    if not paths:
+        return Failure(ValueError('No paths'))
+
+    path, rest = paths[0], paths[1:]
+
+    if not rest:
+        return create_iterable(raw_data, path)
+
+    my_list: list = []
+
+    for iterable in create_iterable(raw_data, path).unwrap():
+
+        iterable_data_handler(iterable, rest).map(
+            my_list.extend,
+        )
+
+    return Success(my_list)
+
+
+def create_iterable(input_data, iterable) -> ResultE[list]:
+    """Return set of set of data per entry in list at iterable[path]."""
+    return fetch_list_by_keys(
+        input_data,
+        iterable[PATH],
+    ).map(
+        lambda collections: [
+            {
+                **input_data,
+                **{iterable[ALIAS]: collection},
+            }
+            for collection in collections
+        ],
+    ).fix(
+        lambda _: [input_data],
+    )