+++ tests/test_mapper.py
@@ -1,3 +1,5 @@
+from returns.pipeline import is_successful
+
 from piri.mapper import map_data
 
 
@@ -49,12 +51,81 @@ def test_array_true_but_no_loop_gives_array():
     ).unwrap() == [{'name': 'test name'}]
 
 
+def test_missing_data_gives_nothing():
+    """Test that we get an array if we set array = true in object."""
+    input_data = {'key': 'test name'}
+    config = {
+        'name': 'root',
+        'array': True,
+        'attributes': [
+            {
+                'name': 'name',
+                'mappings': [
+                    {
+                        'path': ['missing'],
+                    },
+                ],
+            },
+        ],
+    }
+
+    assert not is_successful(map_data(
+        input_data,
+        config,
+    ))
+
+
+def test_missing_data_creates_no_object():
+    """Test that if an object mapping result is empty we create now 'key'."""
+    input_data = {'key': 'test name'}
+    config = {
+        'name': 'root',
+        'array': True,
+        'attributes': [
+            {
+                'name': 'an_attribute',
+                'default': 'val',
+            },
+        ],
+        'objects': [
+            {
+                'name': 'test',
+                'array': False,
+                'attributes': [
+                    {
+                        'name': 'name',
+                        'mappings': [
+                            {
+                                'path': ['missing'],
+                            },
+                        ],
+                    },
+                ],
+            },
+        ],
+    }
+
+    expected_result = [{
+        'an_attribute': 'val',
+    }]
+
+    assert map_data(
+        input_data,
+        config,
+    ).unwrap() == expected_result
+
+
 def test_double_repeatable():
     """Test that we can map nested repeatable objects."""
     config = {
         'name': 'root',
         'array': True,
-        'path_to_iterable': ['journals'],
+        'iterables': [
+            {
+                'alias': 'journals',
+                'path': ['journals'],
+            },
+        ],
         'attributes': [
             {
                 'name': 'journal_id',
@@ -69,8 +140,11 @@ def test_double_repeatable():
             {
                 'name': 'invoices',
                 'array': True,
-                'path_to_iterable': [
-                    'journals', 'journal', 'invoices',
+                'iterables': [
+                    {
+                        'alias': 'invoices',
+                        'path': ['journals', 'journal', 'invoices'],
+                    },
                 ],
                 'attributes': [
                     {
@@ -129,7 +203,12 @@ def test_mapping_where_data_is_not_found():
     config = {
         'name': 'root',
         'array': True,
-        'path_to_iterable': ['journals'],
+        'iterables': [
+            {
+                'alias': 'journals',
+                'path': ['journals'],
+            },
+        ],
         'attributes': [
             {
                 'name': 'journal_id',
@@ -144,8 +223,11 @@ def test_mapping_where_data_is_not_found():
             {
                 'name': 'invoices',
                 'array': True,
-                'path_to_iterable': [
-                    'journals', 'journal', 'invoices',
+                'iterables': [
+                    {
+                        'alias': 'invoices',
+                        'path': ['journals', 'journal', 'invoices'],
+                    },
                 ],
                 'attributes': [
                     {
@@ -202,6 +284,7 @@ def test_mapping_where_data_is_not_found():
         },
         {
             'journal_id': 2,
+            'invoices': [],
         },
     ]
 
@@ -279,7 +362,12 @@ def test_most_features():
             {
                 'name': 'people',
                 'array': True,
-                'path_to_iterable': ['persons'],
+                'iterables': [
+                    {
+                        'alias': 'persons',
+                        'path': ['persons'],
+                    },
+                ],
                 'attributes': [
                     {
                         'name': 'firstname',